event: Add reply-making methods to RoomMessageEventWithoutRelation
